Make Command optionally cancellable
Remove TokenCancellableCommand and merge cancellation functionality into Command.
