+++ src/ReactiveDomain.Messaging/Bus/CommandPublisher.cs
@@ -27,10 +27,9 @@ namespace ReactiveDomain.Messaging.Bus
         }
         public void Fire(Command command, string exceptionMsg = null, TimeSpan? responseTimeout = null, TimeSpan? ackTimeout = null)
         {
-            var tCmd = command as TokenCancellableCommand;
-            if (tCmd?.IsCanceled ?? false)
+            if (command.IsCanceled)
             {
-                _publisher.Publish(tCmd.Canceled());
+                _publisher.Publish(command.Canceled());
                 return;
             }
 
@@ -52,10 +51,9 @@ namespace ReactiveDomain.Messaging.Bus
             {
                 //todo: we're not chaining through the fire method here because it doesn't give 
                 //us the command response to return so there are some duplicated checks 
-                var tCmd = command as TokenCancellableCommand;
-                if (tCmd?.IsCanceled ?? false)
+                if (command.IsCanceled)
                 {
-                    response = tCmd.Canceled();
+                    response = command.Canceled();
                     _publisher.Publish(response);
                     return false;
                 }
